+++ snarkos/config.rs
@@ -111,6 +111,8 @@ pub struct Storage {
     pub trim: bool,
     /// If `true`, scans superfluous blocks for valid forks at boot time. Can take a while.
     pub scan_for_forks: bool,
+    /// If `true`, uses sqlite instead of rocksdb for backend storage. Requires `snarkos_storage/sqlite_storage` feature.
+    pub use_sqlite: bool,
 }
 
 impl Default for Config {
@@ -155,6 +157,7 @@ impl Default for Config {
                 trim: false,
                 validate: false,
                 scan_for_forks: false,
+                use_sqlite: false,
             },
         }
     }
@@ -229,6 +232,7 @@ impl Config {
             "no-jsonrpc" => self.no_jsonrpc(arguments.is_present(option)),
             "trim-storage" => self.trim_storage(arguments.is_present(option)),
             "validate-storage" => self.validate_storage(arguments.is_present(option)),
+            "sqlite" => self.storage.use_sqlite = arguments.is_present(option),
             // Options
             "connect" => self.connect(arguments.value_of(option)),
             "export-canon-blocks" => self.export_canon_blocks(clap::value_t!(arguments.value_of(*option), u32).ok()),
@@ -429,6 +433,7 @@ impl CLI for ConfigCli {
         flag::IS_CRAWLER,
         flag::TRIM_STORAGE,
         flag::VALIDATE_STORAGE,
+        flag::SQLITE,
     ];
     const NAME: NameType = "snarkOS";
     const OPTIONS: &'static [OptionType] = &[
@@ -461,6 +466,7 @@ impl CLI for ConfigCli {
             "import-canon-blocks",
             "is-bootnode",
             "is-miner",
+            "sqlite",
             "is-crawler",
             "ip",
             "port",
