+++ storage/src/storage/mod.rs
@@ -14,25 +14,23 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-#![allow(clippy::needless_lifetimes)]
-
-pub mod exporter;
-pub use exporter::*;
-
-pub mod trim;
-pub use trim::*;
-
+pub mod mem;
+pub use mem::*;
+#[cfg(feature = "rocksdb_storage")]
+mod rocks;
+#[cfg(feature = "rocksdb_storage")]
+pub use rocks::*;
 pub mod storage;
 pub use storage::*;
+pub mod sync;
+pub use sync::*;
+pub mod async_adapter;
+pub use async_adapter::*;
 
-pub mod objects;
-pub use objects::*;
+pub mod key_value;
+pub use key_value::KeyValueStorage;
 
-pub mod digest;
-pub use digest::*;
-
-pub mod validator;
-pub use validator::*;
-
-/// The number of block hashes that are returned by the `Ledger::get_block_locator_hashes` call.
-pub const NUM_LOCATOR_HASHES: u32 = 64;
+#[cfg(feature = "sqlite_storage")]
+pub mod sqlite;
+#[cfg(feature = "sqlite_storage")]
+pub use sqlite::*;
