+++ storage/src/storage/sqlite/mod.rs
@@ -1,0 +1,64 @@
+// Copyright (C) 2019-2021 Aleo Systems Inc.
+// This file is part of the snarkOS library.
+
+// The snarkOS library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// The snarkOS library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
+
+use anyhow::*;
+use rusqlite::{
+    types::{FromSql, FromSqlError, FromSqlResult, ToSqlOutput, ValueRef},
+    Connection,
+    OpenFlags,
+    ToSql,
+};
+use std::path::Path;
+
+use crate::{BlockStatus, Digest, SerialBlock, SerialBlockHeader, TransactionLocation};
+
+mod sync;
+
+pub struct SqliteStorage {
+    conn: Connection,
+}
+
+impl SqliteStorage {
+    pub fn new_ephemeral() -> Result<Self> {
+        Ok(Self {
+            conn: Connection::open_in_memory()?,
+        })
+    }
+
+    pub fn new<P: AsRef<Path>>(path: P) -> Result<Self> {
+        Ok(Self {
+            conn: Connection::open_with_flags(
+                path,
+                OpenFlags::SQLITE_OPEN_READ_WRITE | OpenFlags::SQLITE_OPEN_CREATE | OpenFlags::SQLITE_OPEN_NO_MUTEX,
+            )?,
+        })
+    }
+}
+
+impl FromSql for Digest {
+    fn column_result(value: ValueRef<'_>) -> FromSqlResult<Self> {
+        match value {
+            ValueRef::Blob(blob) => Ok(Self::from(blob)),
+            _ => Err(FromSqlError::InvalidType),
+        }
+    }
+}
+
+impl ToSql for Digest {
+    fn to_sql(&self) -> rusqlite::Result<ToSqlOutput<'_>> {
+        Ok(ToSqlOutput::Borrowed(ValueRef::Blob(&self[..])))
+    }
+}
